Consolidate indicator utilities package
