+++ ai_trading/features/__init__.py
@@ -1,7 +1,8 @@
 """
 Feature engineering public API.
 """
-from .indicators import compute_atr, compute_macd, compute_macds, compute_vwap, ensure_columns
+from ai_trading.indicators import compute_atr
+from .indicators import compute_macd, compute_macds, compute_vwap, ensure_columns
 
 def build_features_pipeline(df, symbol: str):
     """
